requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
python-dotenv==0.19.2
requests==2.27.1
Werkzeug==2.0.3